backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==2.2.5
uuid==1.30
pydantic==1.10.13
pyjwt==2.8.0 
requests==2.32.3
google-generativeai==0.3.2